pyproject.toml
[tool.poetry]
name = "p12-0624"
version = "0.1.0"
description = ""
authors = ["Myryos"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
peewee = "^3.17.6"
typer = "^0.12.3"
sentry-sdk = "^2.7.1"
inquirer = "^3.3.0"
inquirerpy = "^0.3.4"
bcrypt = "^4.1.3"
cryptography = "^42.0.8"
python-dotenv = "^1.0.1"
pyjwt = "^2.8.0"
pytest = "^8.3.2"
pytest-mock = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
